refactor: update route
