+++ src/Campaigns/ServiceProvider.php
@@ -55,7 +55,7 @@ class ServiceProvider implements ServiceProviderInterface
         Hooks::addAction('rest_api_init', Routes\RegisterCampaignRoutes::class);
         Hooks::addAction('rest_api_init', Routes\GetCampaignsListTable::class, 'registerRoute');
         Hooks::addAction('rest_api_init', Routes\DeleteCampaignListTable::class, 'registerRoute');
-        Hooks::addAction('rest_api_init', Routes\UpdateDefaultCampaignFormListTable::class, 'registerRoute', 9);
+        Hooks::addAction('rest_api_init', Routes\UpdateDefaultCampaignForm::class, 'registerRoute', 9);
         Hooks::addAction('rest_api_init', Routes\GetCampaignStatistics::class, 'registerRoute');
     }
